requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
decorator==5.1.1
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.1.3
executing==1.2.0
google==3.0.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
protobuf==4.24.3
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
